some parameter changes
